# doc/openapi.yaml
...
openapi: 3.0.0
info:
  title: Contappa API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: REST API for managing the Contappa service
servers:
  - url: 127.0.0.1
security: []
paths:
...
  "/tables/{tableId}/bills":
    parameters:
      - $ref: "#/components/parameters/tableId"
    post:
      summary: Create a bill
      operationId: createBill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillRequest"
      responses:
        "201":
          description: Bill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: Get all bills
      operationId: getAllBills
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/tables/{tableId}/bills/{billId}":
    parameters:
      - $ref: "#/components/parameters/tableId"
      - $ref: "#/components/parameters/billId"
    get:
      summary: Get bill by ID
      operationId: getBillById
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update bill
      operationId: updateBill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBillRequest"
      responses:
        "200":
          description: Bill updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete bill
      operationId: deleteBill
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/tables/{tableId}/bills/{billId}/split":
    parameters:
      - $ref: "#/components/parameters/tableId"
      - $ref: "#/components/parameters/billId"
    post:
      summary: Split a bill
      operationId: splitBill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SplitBillRequest"
      responses:
        "201":
          description: Bills created from split
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
        "400":
          description: Invalid split
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Bill not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    tableId:
      name: tableId
      in: path
      required: true
      schema:
        type: string
    billId:
      name: billId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Bill:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        tableId:
          type: string
        paid:
          type: boolean
        products:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
    CreateBillRequest:
      type: object
      properties:
        amount:
          type: number
        date:
          type: string
          format: date
        tableId:
          type: string
        products:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
      required:
        - amount
        - date
        - tableId
        - products
    UpdateBillRequest:
      type: object
      properties:
        amount:
          type: number
        date:
          type: string
          format: date
        tableId:
          type: string
        paid:
          type: boolean
        products:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
    SplitBillRequest:
      type: object
      properties:
        splits:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              products:
                type: array
                minItems: 1
                items:
                  type: object
                  properties:
                    productId:
                      type: string
                    quantity:
                      type: integer
...
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string